requirements.txt
django==5.2.10
djangorestframework
django-cors-headers
africastalking
requests
flask
